neko-social-manager/requirements.txt
python-telegram-bot>=20.0
anthropic>=0.25.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
aiofiles>=23.0.0
